requirements.txt
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dnspython==2.8.0
Faker==40.15.0
fonttools==4.62.1
kiwisolver==1.5.0
matplotlib==3.10.9
mysql-connector-python==9.7.0
numpy==2.4.4
packaging==26.2
pandas==3.0.2
pillow==12.2.0
pymongo==4.17.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
seaborn==0.13.2
six==1.17.0
tqdm==4.67.3
tzdata==2026.2
